+++ packages/gatsby-recipes/src/providers/utils/get-graphql-fields.test.js
@@ -17,19 +17,19 @@ test(`get-graphql-fields returns an array of fields`, () => {
   expect(result).toMatchInlineSnapshot(`
     Array [
       Object {
-        "fields": Object {
-          "body": Object {
-            "metadata": Object {
-              "title": "Content",
-              "type": "Text",
-            },
+        "fields": Array [
+          Object {
+            "id": "title",
+            "name": "title",
             "type": "String",
           },
-          "title": Object {
-            "metadata": undefined,
-            "type": "String",
+          Object {
+            "id": "body",
+            "name": "body",
+            "title": "Content",
+            "type": "Text",
           },
-        },
+        ],
         "name": "BlogPost",
       },
     ]