+++ packages/gatsby/src/utils/develop-proxy.ts
@@ -16,10 +16,7 @@ interface IProxyControls {
 
 const noop = (): void => {}
 
-const adminFolder = path.join(
-  path.dirname(require.resolve(`gatsby-admin`)),
-  `public`
-)
+const adminFolder = path.join(__dirname, `..`, `..`, `gatsby-admin-public`)
 
 const serveAdmin = st({
   path: adminFolder,
@@ -64,6 +61,7 @@ export const startDevelopProxy = (input: {
     }
 
     if (req.url === `/socket.io/socket.io.js`) {
+      res.setHeader(`Content-Type`, `application/javascript`)
       res.end(
         fs.readFileSync(require.resolve(`socket.io-client/dist/socket.io.js`))
       )