+++ packages/gatsby-remark-copy-linked-files/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-copy-linked-files@2.3.11...gatsby-remark-copy-linked-files@2.3.12) (2020-07-28)
+
+### Bug Fixes
+
+- do not augment plugin-options ([#26006](https://github.com/gatsbyjs/gatsby/issues/26006)) ([222920b](https://github.com/gatsbyjs/gatsby/commit/222920b))
+
+## [2.3.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-copy-linked-files@2.3.10...gatsby-remark-copy-linked-files@2.3.11) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-remark-copy-linked-files
+
+## [2.3.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-copy-linked-files@2.3.9...gatsby-remark-copy-linked-files@2.3.10) (2020-07-02)
+
+**Note:** Version bump only for package gatsby-remark-copy-linked-files
+
+## [2.3.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-copy-linked-files@2.3.8...gatsby-remark-copy-linked-files@2.3.9) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-remark-copy-linked-files
+
+## [2.3.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-copy-linked-files@2.3.7...gatsby-remark-copy-linked-files@2.3.8) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-remark-copy-linked-files
+
+## [2.3.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-copy-linked-files@2.3.6...gatsby-remark-copy-linked-files@2.3.7) (2020-06-24)
+
+**Note:** Version bump only for package gatsby-remark-copy-linked-files
+
+## [2.3.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-copy-linked-files@2.3.5...gatsby-remark-copy-linked-files@2.3.6) (2020-06-22)
+
+**Note:** Version bump only for package gatsby-remark-copy-linked-files
+
 ## [2.3.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-copy-linked-files@2.3.4...gatsby-remark-copy-linked-files@2.3.5) (2020-06-09)
 
 **Note:** Version bump only for package gatsby-remark-copy-linked-files